zng_wgt_transform
Function backface_visibility

pub fn backface_visibility(
    child: impl UiNode,
    visible: impl IntoVar<bool>,
) -> impl UiNode

P Sets if the widget is still visible when it is turned back towards the viewport due to rotations in X or Y axis in the widget or in parent widgets.

Widget back face is visible by default, the back face is a mirror image of the front face, if visible is set to false the widget is still layout and rendered, but it is not displayed on screen by the view-process if the final global transform of the widget turns the backface towards the viewport.

This property affects any descendant widgets too, unless they also set backface_visibility.