zng_wgt_transform

Function rotate_y

pub fn rotate_y(
    child: impl IntoUiNode,
    angle: impl IntoVar<AngleRadian>,
) -> UiNode

P Rotate transform.

This property is a shorthand way of setting transform to new_rotate_y(angle) using variable mapping.

The rotation is done around the y axis that passes trough the transform_origin in 3D.