Crate zng_wgt_transform

Transform properties, scale, rotate, transform and more.

Crate

This crate is part of the zng project.

Statics

PERSPECTIVE_ORIGIN_VAR

Vanishing point used by transform when it is 3D.

TRANSFORM_ORIGIN_VAR

Point relative to the widget inner bounds around which the transform is applied.

Functions

backface_visibility

P Sets if the widget is still visible when it is turned back towards the viewport due to rotations in X or Y axis in the widget or in parent widgets.

perspective

P Distance from the Z plane (0) the viewer is, affects 3D transform on the widget’s children.

perspective_origin

P Vanishing point used by 3D transforms in the widget’s children.

rotate

P Rotate transform.

rotate_x

P Rotate transform.

rotate_y

P Rotate transform.

rotate_z

P Same as rotate.

scale

P Scale transform.

scale_x

P Scale X transform.

scale_xy

P Scale X and Y transform.

scale_y

P Scale Y transform.

skew

P Skew transform.

skew_x

P Skew X transform.

skew_y

P Skew Y transform.

transform

P Custom transform.

transform_origin

P Point relative to the widget inner bounds around which the transform is applied.

transform_style

P Defines how the widget and children are positioned in 3D space.

translate

P Translate transform.

translate_x

P Translate X transform.

translate_y

P Translate Y transform.

translate_z

P Translate Z transform.